refactor(Money): constrain `_minorUnits` property to `Int64` from `Int`

+++ Sources/SwiftMoney/Money/Money+Arithmetic.swift
@@ -11,11 +11,11 @@ public extension Money {
 }
 
 public extension Money {
-    static func * (lhs: Money, rhs: Int) -> Money {
+    static func * (lhs: Money, rhs: Int64) -> Money {
         Money(minorUnits: lhs._minorUnits * rhs)
     }
 
-    static func * (lhs: Int, rhs: Money) -> Money {
+    static func * (lhs: Int64, rhs: Money) -> Money {
         Money(minorUnits: lhs * rhs._minorUnits)
     }
 }
